Fixed error introduced by adding production env to Maven build
When NODE_ENV is production npm will not install devDependencies which causes the frontend build to fail because webpack is not installed locally. This usually only happens on the first install but can also happen after a clean. Moved install out into its own exec block in pom.xml which now runs prior to the frontend build.